Rename logging.rb file -> exception_handling.rb
